Create Email Generation
Macro generates an email and attaches the file (with only specific tabs so a temp file is created). To, cc, Bcc and Subject are grabbed from a worksheet to give the user control. Temporary file  is deleted once email is generated. MACRO DOES NOT SEND THE EMAIL so as to give the user final control (double checking and adding other content if need be) before sending the email.